Add first test for display work serialisation
