+++ sbt_common/display/src/test/scala/uk/ac/wellcome/display/models/DisplayJacksonModuleTestBase.scala
@@ -1,0 +1,139 @@
+package uk.ac.wellcome.display.models
+
+import io.circe.Json
+import io.circe.parser._
+import org.scalatest.Suite
+import uk.ac.wellcome.models._
+
+trait DisplayJacksonModuleTestBase {this: Suite =>
+
+  def items(its: List[IdentifiedItem]) =
+    its
+      .map { it =>
+        s"""{
+          "id": "${it.canonicalId}",
+          "type": "${it.ontologyType}",
+          "locations": [
+            ${locations(it.locations)}
+          ]
+        }"""
+      }
+      .mkString(",")
+
+  def locations(locations: List[Location]) =
+    locations
+      .map { location(_) }
+      .mkString(",")
+
+  def location(loc: Location) = loc match {
+    case l: DigitalLocation => digitalLocation(l)
+    case l: PhysicalLocation => physicalLocation(l)
+  }
+
+  def digitalLocation(loc: DigitalLocation) =
+    s"""{
+      "type": "${loc.ontologyType}",
+      "locationType": "${loc.locationType}",
+      "url": "${loc.url}",
+      "license": ${license(loc.license)}
+    }"""
+
+  def physicalLocation(loc: PhysicalLocation) =
+    s"""
+       {
+        "type": "${loc.ontologyType}",
+        "locationType": "${loc.locationType}",
+        "label": "${loc.label}"
+       }
+     """
+
+  def license(license: License) =
+    s"""{
+      "label": "${license.label}",
+      "licenseType": "${license.licenseType}",
+      "type": "${license.ontologyType}",
+      "url": "${license.url}"
+    }"""
+
+  def identifier(identifier: SourceIdentifier) =
+    s"""{
+      "type": "Identifier",
+      "identifierScheme": "${identifier.identifierScheme}",
+      "value": "${identifier.value}"
+    }"""
+
+  def identifiedOrUnidentifiable[T](displayableAgent: Displayable[T],
+                                    f: T => String) =
+    displayableAgent match {
+      case Unidentifiable(ag) => f(ag)
+      case Identified(ag, id, identifiers) =>
+        val agent = parse(f(ag)).right.get.asObject.get
+        val identifiersJson = identifiers.map { sourceIdentifier =>
+          parse(identifier(sourceIdentifier)).right.get
+        }
+        val newJson = ("id", Json.fromString(id)) +: (
+          "identifiers",
+          Json.arr(identifiersJson: _*)) +: agent
+        Json.fromJsonObject(newJson).spaces2
+    }
+
+  def abstractAgent(ag: AbstractAgent) =
+    ag match {
+      case a: Agent => agent(a)
+      case o: Organisation => organisation(o)
+      case p: Person => person(p)
+    }
+
+  def person(p: Person) = {
+    s"""{
+        "type": "Person",
+        ${optionalString("prefix", p.prefix)},
+        ${optionalString("numeration", p.numeration)},
+        "label": "${p.label}"
+      }"""
+  }
+
+  def organisation(o: Organisation) = {
+    s"""{
+        "type": "Organisation",
+        "label": "${o.label}"
+      }"""
+  }
+
+  def agent(a: Agent) = {
+    s"""{
+        "type": "Agent",
+        "label": "${a.label}"
+      }"""
+  }
+
+  def optionalString(fieldName: String, maybeValue: Option[String]) =
+    maybeValue match {
+      case None => ""
+      case Some(p) =>
+        s"""
+           "$fieldName": "$p"
+         """
+    }
+
+  def period(p: Period) =
+    s"""{
+      "type": "Period",
+      "label": "${p.label}"
+    }"""
+
+
+  def concept(con: AbstractConcept) =
+    s"""
+    {
+      "type": "${con.ontologyType}",
+      "label": "${con.label}"
+    }
+    """
+
+  def concepts(concepts: List[AbstractConcept]) =
+    concepts
+      .map { concept(_) }
+      .mkString(",")
+
+}
